% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{devdiff}
\alias{devdiff}
\title{Difference in Deviances}
\usage{
devdiff(X, LF_alt, LF_null = NULL)
}
\arguments{
\item{X}{a data matrix.}

\item{LF_alt}{Observed logistic factors.}

\item{LF_null}{Null logistic factors.}
}
\description{
Difference in Deviances
}
\author{
Wei Hao
}
\keyword{internal}
