% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{devdiff_parallel}
\alias{devdiff_parallel}
\title{Difference in Deviances, in Parallel}
\usage{
devdiff_parallel(X, LF_alt, LF_null = NULL, numcores = 1)
}
\arguments{
\item{X}{a data matrix.}

\item{LF_alt}{Observed logistic factors.}

\item{LF_null}{Null logistic factors.}
}
\description{
Difference in Deviances, in Parallel
}
\author{
Wei Hao
}
\keyword{internal}

