% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_k.R
\name{find_k}
\alias{find_k}
\title{Find a number of clusters or principal components}
\usage{
find_k()
}
\description{
There are a wide range of algorithms and visual techniques to identify
a number of clusters or principal components embedded in the observed data.
}
\details{
It is critical to explore the eigenvalues, cluster stability, and visualization.
See R packages \code{bootcluster}, \code{EMCluster}, and \code{nFactors}.

Please see the R package \code{SC3}, which provides \code{estkTW()} function to
find the number of significant eigenvalues according to the Tracy-Widom test.

\code{ADPclust} package includes \code{adpclust()} function that runs the algorithm
on a range of K values. It helps you to identify the most suitable number of clusters.

This package also provides an alternative methods in \code{permutationPA}.
Through a resampling-based Parallel Analysis, it finds a number of significant components.
}
