% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackstraw_alstructure.R
\name{jackstraw_alstructure}
\alias{jackstraw_alstructure}
\title{Non-Parametric Jackstraw for ALStructure}
\usage{
jackstraw_alstructure(
  dat,
  FUN = function(x) t(alstructure::alstructure(x, d_hat = r, svd_method =
    "truncated_svd", tol = 0.001, max_iters = 1000)$Q_hat[, , drop = FALSE]),
  devR = FALSE,
  r = NULL,
  r1 = NULL,
  s = NULL,
  B = NULL,
  covariate = NULL,
  verbose = TRUE,
  seed = NULL
)
}
\arguments{
\item{dat}{a genotype matrix with \code{m} rows as variables and \code{n} columns as observations.}

\item{FUN}{a function to ALStructure}

\item{devR}{use a R function to compute deviance. By default, FALSE (uses C++).}

\item{r}{a number of significant LFs.}

\item{r1}{a numeric vector of LFs of interest (implying you are not interested in all \code{r} LFs).}

\item{s}{a number of ``synthetic'' null variables. Out of \code{m} variables, \code{s} variables are independently permuted.}

\item{B}{a number of resampling iterations. There will be a total of \code{s*B} null statistics.}

\item{covariate}{a data matrix of covariates with corresponding \code{n} observations (do not include an intercept term).}

\item{verbose}{a logical specifying to print the computational progress.}

\item{seed}{a seed for the random number generator.}
}
\value{
\code{jackstraw_alstructure} returns a list consisting of
\item{p.value}{\code{m} p-values of association tests between variables and their LFs}
\item{obs.stat}{\code{m} observed devs}
\item{null.stat}{\code{s*B} null devs}
}
\description{
Test association between the observed variables and population structure estimated by ALStructure.
}
\details{
This function uses ALStructure from Cabreros and Storey (2019). A deviation \code{dev} in logistic regression
(the full model with \code{r} LFs vs. the intercept-only model) is used to assess association.
}
\seealso{
\link{jackstraw_pca} \link{jackstraw}
}
\author{
Neo Christopher Chung \email{nchchung@gmail.com}
}
