% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jackstraw_clara.R
\name{jackstraw_clara}
\alias{jackstraw_clara}
\title{Non-Parametric Jackstraw for CLARA (Clustering Large Applications)}
\usage{
jackstraw_clara(dat, clara.dat, s = NULL, B = NULL, covariate = NULL,
  verbose = FALSE, pool = TRUE, seed = NULL, ...)
}
\arguments{
\item{dat}{a matrix with \code{m} rows as variables and \code{n} columns as observations.}

\item{clara.dat}{an output from applying \code{cluster::clara()} on \code{dat}.}

\item{s}{a number of ``synthetic'' null variables. Out of \code{m} variables, \code{s} variables are independently permuted.}

\item{B}{a number of resampling iterations.}

\item{covariate}{a model matrix of covariates with \code{n} observations. Must include an intercept in the first column.}

\item{verbose}{a logical specifying to print the computational progress. By default, \code{FALSE}.}

\item{pool}{a logical specifying to pool the null statistics across all clusters. By default, \code{TRUE}.}

\item{seed}{a seed for the random number generator.}

\item{...}{optional arguments to control the k-means clustering algorithm (refers to \code{kmeans}).}
}
\value{
\code{jackstraw_clara} returns a list consisting of
\item{F.obs}{\code{m} observed F statistics between variables and cluster medoids.}
\item{F.null}{F null statistics between null variables and cluster medoids, from the jackstraw method.}
\item{p.F}{\code{m} p-values of membership.}
}
\description{
Test the cluster membership for CLARA (Clustering Large Applications)
}
\details{
CLARA assigns \code{m} rows into \code{K} clusters. This function enable statistical
evaluation if the cluster membership is correctly assigned. Each of \code{m} p-values refers to
the statistical test of that row with regard to its assigned cluster.
Its resampling strategy accounts for the over-fitting characteristics due to direct computation of clusters from the observed data
and protects against an anti-conservative bias.

The input data (\code{dat}) must be of a class `matrix`.

Note that \code{cluster::clara()} is ran with the following options arguments,
\code{samples = floor(nrow(dat)/100), sampsize = floor(nrow(dat)/100), pamLike=TRUE}.
They dictate the accuracy and speed of CLARA, see \code{?cluster::clara()}.
}
\examples{
set.seed(1234)
dat = t(scale(t(Jurkat293T), center=TRUE, scale=FALSE))
clara.dat <- clara(dat, k=2)
jackstraw.out <- jackstraw_clara(dat, clara.dat = clara.dat)
}
\references{
Chung (2018) Statistical significance for cluster membership. biorxiv, doi:10.1101/248633 \url{https://www.biorxiv.org/content/early/2018/01/16/248633}
}
\author{
Neo Christopher Chung \email{nchchung@gmail.com}
}
