% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jurkat293T.R
\docType{data}
\name{Jurkat293T}
\alias{Jurkat293T}
\title{A Jurkat:293T equal mixture dataset from Zheng et al. (2017)}
\format{
A data frame with 3381 rows corresponding to single cells and 10 columns corresponding to the top 10 principal components
}
\source{
Supplementary Data 1 from Zheng et al. (2017) \url{https://static-content.springer.com/esm/art\%3A10.1038\%2Fncomms14049/MediaObjects/41467_2017_BFncomms14049_MOESM829_ESM.xlsx}
}
\usage{
Jurkat293T
}
\description{
50%:50% Jurkat:293T
}
\references{
Zheng et al. (2017) Massively parallel digital transcriptional profiling of single cells. Nature Communications. 8:14049. DOI: 10.1038/ncomms14049
}
\keyword{datasets}
