% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi0est_bootstrap.R
\name{pi0est_bootstrap}
\alias{pi0est_bootstrap}
\title{Pi0 estimation using a non-analytical bootstrap algorithm}
\usage{
pi0est_bootstrap(p = NULL, lambda = seq(0, 0.9, 0.05), robust = FALSE)
}
\arguments{
\item{lambda}{the value of the tuning parameter to estimate pi0 (optional)}

\item{robust}{an indicator of whether it is desired to make the estimate more robust for small p-values and a direct finite sample estimate of pFDR (optional)}

\item{pvalue}{a vector of p-values.}

\item{verbose}{print a progress}

\item{...}{optional arguments to control a local FDR estimation.}
}
\value{
\code{pi0est_bootstrap} returns a pi0 estimate
}
\description{
This is a function to estimate pi0 using the bootstrap, directly dervied from the qvalue version 1.43 (Bioconductor 3.0).
Try \code{qvalue::pi0est} or other pi0 estimation methods. This function is preserved because in the newer qvalue package (version > 2.0), the bootstrap method is replaced by a closed form solution.
In practice, the closed form solution may fail due to a very small pi0 (e.g., generally small p-values) or others.
}
\author{
John D. Storey, Alan Dabney
}
