% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pip.R
\name{pip}
\alias{pip}
\title{Compute posterior inclusion probabilities (PIPs)}
\usage{
pip(pvalue, group = NULL, pi0 = NULL, ...)
}
\arguments{
\item{pvalue}{a vector of p-values.}

\item{group}{a vector of group indicators (optional).}

\item{pi0}{a vector of pi0 values (optional)}

\item{...}{optional arguments for \code{\link[qvalue]{lfdr}} to control a local FDR estimation.}
}
\value{
\code{pip} returns a vector of posterior inclusion probabilities
}
\description{
Computes posterior probabilities that a feature is a true member of an assigned cluster
}
\author{
Neo Christopher Chung \email{nchchung@gmail.com}
}
