% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{pseudo_Rsq}
\alias{pseudo_Rsq}
\title{Mcfadden's Pseudo R-squared}
\usage{
pseudo_Rsq(X, LF_alt, LF_null = NULL)
}
\arguments{
\item{X}{a data matrix.}

\item{LF_alt}{Observed logistic factors.}

\item{LF_null}{Null logistic factors.}
}
\description{
Mcfadden's Pseudo R-squared
}
\author{
Wei Hao
}
\keyword{internal}
