% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{efron_Rsq}
\alias{efron_Rsq}
\title{Efron's Pseudo R-squared}
\usage{
efron_Rsq(X, LF)
}
\arguments{
\item{X}{a data matrix.}

\item{LF}{Observed logistic factors.}
}
\description{
Efron's Pseudo R-squared
}
\author{
Wei Hao
}
\keyword{internal}
